requirements-dev.txt
# Development dependencies for PyForge CLI
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# Code formatting and linting
black>=23.0.0
ruff>=0.1.0
isort>=5.12.0

# Type checking
mypy>=1.0.0
types-requests

# Security scanning
bandit[toml]>=1.7.5
safety>=2.3.0

# Build and publishing
build>=0.10.0
twine>=4.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocs-mermaid2-plugin>=0.6.0

# Development utilities
ipython>=8.0.0
jupyter>=1.0.0

# Big Data Processing (for Databricks extension tests)
# PySpark 3.5.0 is used by Databricks Runtime 14.3 LTS (Serverless V1)
pyspark==3.5.0
delta-spark==3.1.0